requirements.txt
# requirements.txt

absl-py==2.1.0
alembic==1.13.2
aniso8601==9.0.1
asttokens==2.4.1
attrs==24.2.0
blinker==1.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dnspython==2.6.1
email-validator==1.1.3
exceptiongroup==1.2.2
executing==2.1.0
Flask==2.0.2
Flask-Dance==5.1.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flatbuffers==24.3.25
fonttools==4.53.1
greenlet==3.0.3
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jax==0.4.33
jaxlib==0.4.33
jedi==0.19.1
Jinja2==3.0.3
jsmin==3.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lesscpy==0.15.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml_dtypes==0.5.0
nest-asyncio==1.6.0
numpy==2.1.1
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
ply==3.11
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
pywin32==306
pyzmq==26.2.0
rcssmin==1.1.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
scipy==1.14.1
six==1.16.0
sounddevice==0.5.0
SQLAlchemy==1.4.29
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
URLObject==2.4.3
wcwidth==0.2.13
Werkzeug==2.0.3
WTForms==3.0.1
xxhash==3.5.0
